lib.songs.auto_find_song: Return the first song if nothing matches mood.
The function was not doing what was advertised.  It would return None if
there were songs in  the user's queue but None fit  the mood.  It should
instead  return the  first  song  in the  user's  queue  (for the  given
station).  There were no tests to catch this so tests were added.