songs.scrobble_song: handle TimeoutErrors
Or OSErrors in general.  This assumes it's a temporary error and logs in again.
Later I hope to re-write the scrobbling system to be a seperate thread with a
Queue on it where errors just cause a sleep/retry and the time between sleeps
increments upon susequent fails.   But for now I'm lazy.