Add Queue.auto_fill_mood() method.
The Queue has a new auto_fill_* method, auto_fill_mood() which, will select
songs from the library based on the current "mood". It does this by getting the
"mood" for the station in the last hour, then for the first 250 (hardcoded)
artists it sees if there is a SongFile for that artist for that list and if so
returns that SongFile in the list, up to the *entries_needed* parameter.