Back out changesets 9dd3e2b1c743 & 70c5a3c47438.
SimpleLazyObject model instances at the global module level.

While these worked fine in this branch they failed horribly in another branch.
The reason being that the python/django discover runner goes through each
object in the modules and checks to see if they are instances of TestCase, well
when it comes accross SimplLazyObjects, isinstance(obj, TestCase) evaluates the
wrapped function, which is not good when the test runner hasn't set up the
databases yet.  I'm not even sure why it works fine here but not in the other
branch, but it's too much trouble to fix and would basically amount to me
backing it out anyway so I'm backing it out.  The class methods have been
restored.

SimpleLazyObject turns out to be not-so-useful.