auto_fill_mood: Modify queryset instead of creating a new one.
This fixes a big oops where there's supposted to be a chain of queryset
modifications, but at the second point in the chain it's actually re-creating
the queryset, thereby igoring the previous change.

I could not figure out why it kept picking the same artists even though their
"mood" had already expired in the configured timeframe.  It's because it was
throwing out the statement where it filterd by timeframe :|