ui: Replace pnotify with Notifier.js
The pnotify library appears to not be maintained anymore, which actually isn't
a problem but the company that released it has come and gone and their web
site now hosts scamware.  I don't want to link to that site anymore and so
decided to remove it completely.  It's been replaced by Notifier.js.  It has a
simpler API but some of the simplicity is actually welcomed.