async: Don't log_mood() for DJ Ango
This bug was introduced in cs bbe572959743 when I started using mpd's stickers
to get the player id of the song.  MPD's stickers are always strings and so
comparing it to an int (primary key) will always be false.

This change takes the sticker and does a Player.objects.get() (which does an
int() on the string).  The reason why I don't just do the int() myself is
because I also wanted to revert back to the original log message, which
requires the Player object.