songs: New function: split_tag_into_words()
This function takes a mixed-case word and splits it into multiple words.  From
the docstring:

    The best way to to describe this is to show examples:

        "loveSongs" -> "love songs"

        "LoveSongs" -> "love songs"

        "lovesongs" -> "lovesongs"

        "LOVESONGS" -> "lovesongs"

        "love_songs" -> "love songs"

This will be needed because Lastfm uses multi-word tags, but I want to use
hash tags a-la Twitter, which can not be split into words.  This function
takes advantage of mixed case (and the underscore) to help it along.

It's probably inefficient and buggy but works for now and will be improved
over time.