models.Queue: New auto-filler: auto_fill_from_tags()
This is the first auto-filler not targeted for the Main Station.  It will
allow user-stations to put #tags in the station name and have DJ Ango pick
songs from artists with those tags from the TeamPlayer library and spin those
tracks on the user's station.

The code to actually make DJ Ango do this is forthcoming.