DJ Ango fills his queue on user stations with hash tags in the name
Provided settings.SHAKE_THINGS_UP is true.  This allows a user to have a
channel with, say "#PopPunk" in the name and DJ Ango will fill his queue with
pop punk artists.

If the user changes the name and the new name has a hash tag in it, then DJ
Ango will clear his queue of the old songs and add new songs based on the new
name.