auto_fill: Don't fill user stations w/o a hashtag will the entire library
This was a pretty bad bug (with an easy fix).  If the user station didn't have
a hash tag then the if/elif chain falls to the bottom and it fills the channel
with the entire query set.

With this change if the user station doesn't have a hashtag then auto_fill
returns immediately.

auto_fill() needs tests...