mpc.currently_playing: Return "not_playing" if mpd.status() returns no song
I got the following error in a test instance:

[2015-04-13 01:22:25] teamplayer Station 1: Waiting 0 seconds
Exception in thread Thread-1:
Traceback (most recent call last):
  File "/usr/lib64/python3.4/threading.py", line 920, in _bootstrap_inner
    self.run()
  File "/opt/teamplayer/lib/python3.4/site-packages/teamplayer/lib/async.py", line 163, in run
    stickers=['dj', 'player_id'])
  File "/opt/teamplayer/lib/python3.4/site-packages/teamplayer/lib/mpc.py", line 168, in currently_playing
    elapsed_time, total_time = (int(i) for i in status['time'].split(':'))
KeyError: 'time'

Apparently, "currentsong" succeeded but by the time "status" was called the
0-second song had ended.  We shouldn't be playing 0-second songs (that fix
will have to come later) but nevertheless we should account for when it does
occur.

This patch returns the "not_playing" dict when "status" fails to return the
expected song info.