recycle user's removed stations
In this changeset, when a user requests to have their station removed, we
don't actually remove it but set enabled=False.  Then, when the user goes back
to create a station, if they have a disabled station then we just re-enable
it.

This is good so we don't have to re-create a station every time, we can just
re-start the mpd.  Also the user gets a quasi-permalink for his/her station.