docker: entrypoint.sh: wait a bit before calling migrate
Sometimes, on my machine, the database container isn't up and listening before
the wsgi container tries to connect to it.  This usually happens the first
time the containers are started.  This change puts a small sleep before the
wsgi container calls migrate in hopes that the database is ready by then.