docker: introduce editor mode
The compose file now passes the DJANGO_DEBUG environment variable to the
container and bind-mounts the teamplayer source to the container in
/usr/src/teamplayer.  If DJANGO_DEBUG is defined and not equal to "0", the
container instead uses the the teamplayer from /usr/src/teamplayer and runs
the development server instead of gunicorn.  This effectively allows for
changes in the source to become available immediately in the container and
utilizes the develpment server's auto-reload features.