views: about: fix crash when playlog is empty
When the PlayLog doesn't have any entries, then the `.first` method fails
and the view blows up.  This change first checks to see if there are any
entries in the log before calling first.  If there are no entries then the
template simply displays that nothing was played (without the timestamp).