mpc: add a .get_version() method
This commit adds a .get_version() classmethod to the MPC class.  It uses the
Main Station instance to get the version of the running MPD service, so the
Main Station must be running, but you don't/shouldn't need to instantiate it
as theoretically it's always running and all MPC instances are connecting to
the same server version.